TG-674 | Renew fabrick recreate (#251)
* feat: expose getRenewTransactionsByPolicyByUid

make it available for other functions to use this functionality, with the first use case being in payment

* refactor: dynamic collection for transactions save

with the added use case of renew transactions, the function signature was updated in a way to enable its use by different scenarios

* feat: renew fabrick recreate

add to the existing endpoint the possibility of recreating a fabrick integration for a policy that is on the draft renew state. The only difference is the collection from which the data is retrieved. All operations remain the same